+++ cryptol-remote-api/python/cryptol/single_connection.py
@@ -242,3 +242,9 @@ def interrupt() -> None:
 def logging(on : bool, *, dest : TextIO = sys.stderr) -> None:
     """Whether to log received and transmitted JSON."""
     __get_designated_connection().logging(on=on,dest=dest)
+
+def file_deps(m : str, timeout:Optional[float] = None) -> Any:
+    """Get information about a loaded module."""
+    return __get_designated_connection().file_deps(m,timeout=timeout)
+
+